+++ User/bsp/c/fric.c
@@ -2,7 +2,7 @@
 #include "bsp\fric.h"
 #include "bsp\delay.h"
 
-#include "tim.h"
+#include <tim.h>
 
 #include "component\user_math.h"
 
@@ -19,7 +19,7 @@ int8_t BSP_Fric_Start(void) {
 	return 0;
 }
 int8_t BSP_Fric_Set(float duty_cycle) {
-	uint16_t pulse = duty_cycle * UINT16_MAX;
+	uint16_t pulse = (uint16_t)(duty_cycle * (float)UINT16_MAX);
 	__HAL_TIM_SET_COMPARE(&htim1, TIM_CHANNEL_1, pulse);
 	__HAL_TIM_SET_COMPARE(&htim1, TIM_CHANNEL_4, pulse);
 	return 0;
