+++ User/bsp/c/laser.c
@@ -1,7 +1,7 @@
 /* Includes ------------------------------------------------------------------*/
 #include "bsp\laser.h"
 
-#include "tim.h"
+#include <tim.h>
 
 #include "component\user_math.h"
 
@@ -20,7 +20,7 @@ int8_t BSP_Laser_Set(float duty_cycle) {
 	if (duty_cycle > 1.f)
 		return -1;
 	
-	uint16_t pulse = duty_cycle * UINT16_MAX;
+	uint16_t pulse = (uint16_t)(duty_cycle * (float)UINT16_MAX);
 
 	__HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_3, pulse);
